Added some packeges and a few thing
