+++ src/main.py
@@ -1,17 +1,6 @@
 import argparse
 
-
-def get_macros(weight):
-    macros = {}
-    prot = 2.2 * weight
-    macros["Protein"] = prot
-    cal = 40.0 * weight
-    macros["Calories"] = cal
-    macros["Fat"] = cal * 0.3
-    macros["Carbohydrates"] = (0.7 * cal - prot * 4.0) / 4.0
-
-    return macros
-
+from Utility.macros import get_macros
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
@@ -21,4 +10,4 @@ if __name__ == "__main__":
         weight = int(args.weight)
         print(get_macros(weight))
     except ValueError:
-        print("Please enter an integer!")
+        print("Please enter a number!")
